# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  api.ApiResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  api.acceptFriendRequestReq:
    properties:
      request_id:
        type: integer
    required:
    - request_id
    type: object
  api.acceptFriendRequestResp:
    properties:
      friend_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  api.blockUserReq:
    properties:
      blocked_id:
        type: integer
    required:
    - blocked_id
    type: object
...
  api.rejectFriendRequestReq:
    properties:
      request_id:
        type: integer
    required:
    - request_id
    type: object
...
  api.sendFriendRequestReq:
    properties:
      message:
        type: string
      to_user_id:
        type: integer
    required:
    - to_user_id
    type: object
  api.sendFriendRequestResp:
    properties:
      from_user_id:
        type: integer
      request_id:
        type: integer
      status:
        type: integer
      to_user_id:
        type: integer
    type: object
  api.unblockUserReq:
    properties:
      blocked_id:
        type: integer
    required:
    - blocked_id
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: GoIM 即时通讯系统 API — 支持私聊、群聊、朋友圈动态。
  title: GoIM API
  version: "1.0"
paths:
...
  /friend/{friendID}:
    delete:
      consumes:
      - application/json
      description: 删除指定ID的好友关系
      parameters:
      - description: 好友ID
        format: int64
        in: path
        name: friendID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 删除好友
      tags:
      - 好友
  /friend/accept:
    post:
      consumes:
      - application/json
      description: 接受指定ID的好友申请
      parameters:
      - description: 接受信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.acceptFriendRequestReq'
      produces:
      - application/json
      responses:
        "200":
          description: 接受成功
          schema:
            allOf:
            - $ref: '#/definitions/api.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/api.acceptFriendRequestResp'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "403":
          description: 无权操作
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "404":
          description: 申请不存在
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 接受好友申请
      tags:
      - 好友
  /friend/block:
    post:
      consumes:
      - application/json
      description: 将指定用户加入黑名单
      parameters:
      - description: 拉黑信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.blockUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: 拉黑成功
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "409":
          description: 已拉黑
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 拉黑用户
      tags:
      - 好友
  /friend/list:
    get:
      consumes:
      - application/json
      description: 获取当前用户的好友列表
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/api.ApiResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 获取好友列表
      tags:
      - 好友
  /friend/reject:
    post:
      consumes:
      - application/json
      description: 拒绝指定ID的好友申请
      parameters:
      - description: 拒绝信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.rejectFriendRequestReq'
      produces:
      - application/json
      responses:
        "200":
          description: 拒绝成功
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "403":
          description: 无权操作
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "404":
          description: 申请不存在
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 拒绝好友申请
      tags:
      - 好友
  /friend/request:
    post:
      consumes:
      - application/json
      description: 向指定用户发送好友申请
      parameters:
      - description: 申请信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.sendFriendRequestReq'
      produces:
      - application/json
      responses:
        "201":
          description: 申请发送成功
          schema:
            allOf:
            - $ref: '#/definitions/api.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/api.sendFriendRequestResp'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "409":
          description: 已是好友或重复申请
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 发送好友申请
      tags:
      - 好友
  /friend/requests:
    get:
      consumes:
      - application/json
      description: 获取当前用户的待处理好友申请列表
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/api.ApiResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 获取好友申请列表
      tags:
      - 好友
  /friend/unblock:
    post:
      consumes:
      - application/json
      description: 将指定用户从黑名单中移除
      parameters:
      - description: 取消拉黑信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.unblockUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: 取消拉黑成功
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 取消拉黑
      tags:
      - 好友
...
securityDefinitions:
  BearerAuth:
    description: 在 Authorization 头中携带 JWT 访问令牌，格式：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"